pyproject.toml
[tool.poetry]
name = "cloudformation-stack-destroyer"
version = "0.1.0"
description = "Automated development environment stack destroyer"
authors = ["John Watson <https://github.com/johnw-dev>"]
license = "GNU GENERAL PUBLIC LICENSE V3"
readme = "README.md"
packages = [{include = "function"}]

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.73.0"
boto3 = "^1.26.112"

[tool.poetry.group.test.dependencies]
pytest = "^7.3.0"
bandit = "^1.7.5"
safety = "^2.3.5"
black = "^23.1.0"
mock = "^5.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
